Add test function which tests whether required elements are found in a predefined order
The test function allows for elements which are in the given set, but are not expected (i.e. required to be present) to be skipped.

Useful for example if you have a (big) list of events, some of which you would want to compare, and some of which you don't care about and don't want to compare.